+++ src/features/Selection/index.tsx
@@ -12,10 +12,7 @@ const apiRoute = "https://svc.metrotransit.org/nextripv2"
 
 export function Selection(){
     const [fetchedRoutes, setFetchedRoutes] = useState<fieldInput[] | []>([]);
-
     const [selectedRoute, setSelectedRoute] = useState<string>("");
-    const [selectedDirection, setSelectedDirection] = useState<string | number>("");
-    const [selectedStop, setSelectedStop] = useState<string>("");
     const [tableData, setTableData] = useState<any>();
 
     const params = useParams();
@@ -60,9 +57,8 @@ export function Selection(){
             {params.routeId ? 
                 <DirectionSelect id="direction-select" /> : null
             }
-            {(params.routeId || selectedRoute) && (params.directionId || selectedDirection) ? 
+            {(params.routeId || selectedRoute) && (params.directionId) ? 
                 <StopSelect
-                    setSelectedStop={setSelectedStop}
                     id="stop-select"
                 /> : null
             }
